+++ bevy_lint/src/lints/nursery/mod.rs
@@ -1,0 +1,6 @@
+//! Unstable lints that may be removed at any time for any reason.
+//!
+//! For more information, please see [`NURSERY`](crate::groups::NURSERY).
+
+pub mod duplicate_bevy_dependencies;
+pub mod zst_query;