+++ bevy_lint/src/lints/suspicious/insert_unit_bundle.rs
@@ -88,9 +88,11 @@ impl<'tcx> LateLintPass<'tcx> for InsertUnitBundle {
 
         let src_ty = cx.typeck_results().expr_ty(receiver).peel_refs();
 
-        // If the method call was not to `Commands::spawn()` we skip it.
-        if !(match_type(cx, src_ty, &crate::paths::COMMANDS)
-            && method_path.ident.name == self.spawn)
+        // If the method call was not to `Commands::spawn()` or originates from an external macro,
+        // we skip it.
+        if !(span.in_external_macro(cx.tcx.sess.source_map())
+            || match_type(cx, src_ty, &crate::paths::COMMANDS)
+                && method_path.ident.name == self.spawn)
         {
             return;
         }