+++ bevy_lint/CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 **All Changes**: [`lint-v0.2.0...main`](https://github.com/TheBevyFlock/bevy_cli/compare/lint-v0.2.0...main)
 
+### Added
+
+- Lint `iter_current_update_events` to `suspicious` ([#314](https://github.com/TheBevyFlock/bevy_cli/pull/314))
+
 ## [v0.2.0] - 2025-03-19
 
 **All Changes**: [`lint-v0.1.0...lint-v0.2.0`](https://github.com/TheBevyFlock/bevy_cli/compare/lint-v0.1.0...lint-v0.2.0)